React Renderer: refactor more of the js out.
